# docs/generated/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BNPI PMS API",
    "version": "1.0.0",
    "description": "Welcome to the BNPI PMS API documentation. This REST API powers BNPI PMS backend services.",
    "contact": {
      "name": "BNPI PMS Support",
      "email": "support@bnpi-pms.example",
      "url": "https://bnpi-pms.example/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://bnpi-pms.example/terms"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Development server"
    },
    {
      "url": "https://api.bnpi-pms.example/api",
      "description": "Production server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for authentication"
      }
    },
...
  },
  "paths": {
...
    "/api/estimation/{id}": {
      "get": {
        "summary": "Get estimation by ID",
        "description": "Retrieve a specific estimation by its unique identifier with optional field selection",
        "tags": [
          "Estimation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "Estimation ID (MongoDB ObjectId format)",
            "example": "507f1f77bcf86cd799439011"
          },
          {
            "in": "query",
            "name": "fields",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of fields to include (supports nested fields)",
            "example": "id,estimationNumber,estimatedCost,status,projectItems(itemName,estimatedTotal)"
          }
        ],
        "responses": {
          "200": {
            "description": "Estimation retrieved successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Estimation not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "summary": "Update estimation",
        "description": "Update estimation data by ID (partial update)",
        "tags": [
          "Estimation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "Estimation ID",
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "properties": {
                  "estimatedCost": {
                    "type": "number"
                  },
                  "actualCost": {
                    "type": "number",
                    "nullable": true
                  },
                  "marginPercentage": {
                    "type": "number"
                  },
                  "marginAmount": {
                    "type": "number"
                  },
                  "projectedWithMargin": {
                    "type": "number"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "DRAFT",
                      "PENDING",
                      "APPROVED",
                      "REJECTED",
                      "REVISED"
                    ]
                  },
                  "notes": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estimation updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Estimation not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete estimation",
        "description": "Permanently delete an estimation by ID",
        "tags": [
          "Estimation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "Estimation ID",
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "responses": {
          "200": {
            "description": "Estimation deleted successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Estimation not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/estimation": {
      "get": {
        "summary": "Get all estimations",
        "description": "Retrieve estimations with advanced filtering, pagination, sorting, field selection, and optional grouping",
        "tags": [
          "Estimation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Number of records per page"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "description": "Sort order"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Field to sort by",
            "example": "createdAt"
          },
          {
            "in": "query",
            "name": "fields",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of fields to include",
            "example": "id,estimationNumber,status,estimatedCost"
          },
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search query",
            "example": "EST-2025"
          },
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string"
            },
            "description": "JSON array of filter objects",
            "example": "[{\"status\":\"APPROVED\"}]"
          },
          {
            "in": "query",
            "name": "groupBy",
            "schema": {
              "type": "string"
            },
            "description": "Group results by field",
            "example": "status"
          },
          {
            "in": "query",
            "name": "document",
            "schema": {
              "type": "string",
              "enum": [
                "true"
              ]
            },
            "description": "Include estimation documents"
          },
          {
            "in": "query",
            "name": "pagination",
            "schema": {
              "type": "string",
              "enum": [
                "true"
              ]
            },
            "description": "Include pagination metadata"
          },
          {
            "in": "query",
            "name": "count",
            "schema": {
              "type": "string",
              "enum": [
                "true"
              ]
            },
            "description": "Include total count"
          }
        ],
        "responses": {
          "200": {
            "description": "Estimations retrieved successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create new estimation",
        "description": "Create a new estimation with the provided data",
        "tags": [
          "Estimation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "estimationNumber",
                  "projectId",
                  "estimatedCost",
                  "marginPercentage",
                  "marginAmount",
                  "projectedWithMargin"
                ],
                "properties": {
                  "estimationNumber": {
                    "type": "string",
                    "example": "EST-2025-001"
                  },
                  "projectId": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                  },
                  "estimatedCost": {
                    "type": "number",
                    "example": 100000
                  },
                  "actualCost": {
                    "type": "number",
                    "nullable": true,
                    "example": 95000
                  },
                  "marginPercentage": {
                    "type": "number",
                    "example": 30
                  },
                  "marginAmount": {
                    "type": "number",
                    "example": 30000
                  },
                  "projectedWithMargin": {
                    "type": "number",
                    "example": 130000
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "DRAFT",
                      "PENDING",
                      "APPROVED",
                      "REJECTED",
                      "REVISED"
                    ],
                    "default": "DRAFT"
                  },
                  "notes": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Estimation created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}